+++ src/main/java/com/example/moizaspringserver/domain/category/entity/FeedCategory.java
@@ -1,7 +1,5 @@
-package com.example.moizaspringserver.domain.feed.entity;
+package com.example.moizaspringserver.domain.category.entity;
 
-import com.example.moizaspringserver.domain.user.entity.User;
-import com.example.moizaspringserver.global.entity.BaseTimeEntity;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -9,6 +7,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.MapsId;
 import javax.persistence.Table;
+
+import com.example.moizaspringserver.domain.feed.entity.Feed;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,27 +17,27 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "tbl_feed_like")
-public class FeedLike extends BaseTimeEntity {
+@Table(name = "tbl_feed_category")
+public class FeedCategory {
 
 	@EmbeddedId
-	private FeedLikeId id;
-
-	@MapsId("user")
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
+	private FeedCategoryId id;
 
 	@MapsId("feed")
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "feed_id", nullable = false)
 	private Feed feed;
 
+	@MapsId("category")
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "category_id", nullable = false)
+	private Category category;
+
 	@Builder
-	public FeedLike(FeedLikeId id, User user,
-		Feed feed) {
+	public FeedCategory(FeedCategoryId id, Feed feed,
+		Category category) {
 		this.id = id;
-		this.user = user;
 		this.feed = feed;
+		this.category = category;
 	}
 }
