+++ src/main/java/com/example/moizaspringserver/domain/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.example.moizaspringserver.domain.feed.respository;
+package com.example.moizaspringserver.domain.category.repository;
 
-import com.example.moizaspringserver.domain.feed.entity.Category;
+import com.example.moizaspringserver.domain.category.entity.Category;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CategoryRepository extends CrudRepository<Category, Integer> {
