+++ src/main/java/com/example/moizaspringserver/domain/category/repository/FeedCategoryRepository.java
@@ -1,0 +1,9 @@
+package com.example.moizaspringserver.domain.category.repository;
+
+import com.example.moizaspringserver.domain.category.entity.FeedCategory;
+import com.example.moizaspringserver.domain.category.entity.FeedCategoryId;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FeedCategoryRepository extends CrudRepository<FeedCategory, FeedCategoryId> {
+
+}
