--- src/main/java/com/example/moizaspringserver/domain/feed/respository/FeedLikeRepository.java
@@ -1,9 +1,0 @@
-package com.example.moizaspringserver.domain.feed.respository;
-
-import com.example.moizaspringserver.domain.feed.entity.FeedLike;
-import com.example.moizaspringserver.domain.feed.entity.FeedLikeId;
-import org.springframework.data.repository.CrudRepository;
-
-public interface FeedLikeRepository extends CrudRepository<FeedLike, FeedLikeId> {
-
-}
