+++ src/main/java/com/example/moizaspringserver/domain/graduate/entity/GraduateVerifyingFile.java
@@ -1,5 +1,6 @@
-package com.example.moizaspringserver.domain.user.entity;
+package com.example.moizaspringserver.domain.graduate.entity;
 
+import com.example.moizaspringserver.domain.user.entity.User;
 import com.example.moizaspringserver.global.entity.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -7,7 +8,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.Id;
