+++ src/main/java/com/example/moizaspringserver/domain/graduate/repository/GraduateVerifyingFileRepository.java
@@ -1,6 +1,6 @@
-package com.example.moizaspringserver.domain.user.repository;
+package com.example.moizaspringserver.domain.graduate.repository;
 
-import com.example.moizaspringserver.domain.user.entity.GraduateVerifyingFile;
+import com.example.moizaspringserver.domain.graduate.entity.GraduateVerifyingFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GraduateVerifyingFileRepository extends JpaRepository<GraduateVerifyingFile, Integer> {
