+++ src/main/java/com/example/moizaspringserver/domain/like/entity/FeedLike.java
@@ -1,5 +1,8 @@
-package com.example.moizaspringserver.domain.feed.entity;
+package com.example.moizaspringserver.domain.like.entity;
 
+import com.example.moizaspringserver.domain.feed.entity.Feed;
+import com.example.moizaspringserver.domain.user.entity.User;
+import com.example.moizaspringserver.global.entity.BaseTimeEntity;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -15,27 +18,27 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "tbl_feed_category")
-public class FeedCategory {
+@Table(name = "tbl_feed_like")
+public class FeedLike extends BaseTimeEntity {
 
 	@EmbeddedId
-	private FeedCategoryId id;
+	private FeedLikeId id;
+
+	@MapsId("user")
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", nullable = false)
+	private User user;
 
 	@MapsId("feed")
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "feed_id", nullable = false)
 	private Feed feed;
 
-	@MapsId("category")
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "category_id", nullable = false)
-	private Category category;
-
 	@Builder
-	public FeedCategory(FeedCategoryId id, Feed feed,
-		Category category) {
+	public FeedLike(FeedLikeId id, User user,
+		Feed feed) {
 		this.id = id;
+		this.user = user;
 		this.feed = feed;
-		this.category = category;
 	}
 }
