+++ src/main/java/com/example/moizaspringserver/domain/like/repository/FeedLikeRepository.java
@@ -1,0 +1,9 @@
+package com.example.moizaspringserver.domain.like.repository;
+
+import com.example.moizaspringserver.domain.like.entity.FeedLike;
+import com.example.moizaspringserver.domain.like.entity.FeedLikeId;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FeedLikeRepository extends CrudRepository<FeedLike, FeedLikeId> {
+
+}
